Somme la colonne quantité de la coupe au lieu de compter 1 pièce/ligne
Une même clé commande+ligne peut être débitée en plusieurs lots
(nesting sur plusieurs panneaux) : Cutrite génère alors plusieurs
lignes "TC" pour cette clé, chacune portant sa propre quantité de
pièces. Compter "1 par ligne" sous-évaluait donc fortement les
quantités réelles dès qu'une pièce était nestée en plusieurs fois
(ex. 3 lignes à 6+2+1 pièces comptées comme 3 pièces au lieu de 9).

Détecte désormais la colonne "quantité" par comparaison avec les
quantités attendues du fichier de lancement (comme les autres
colonnes, par le contenu, avec un seuil de confiance pour ne pas
se déclencher sur un fichier qui n'a pas cette colonne) et somme
cette quantité par clé au lieu d'incrémenter de 1 par ligne trouvée.

Sur le jeu de données réel : 209 → 11 → 7 anomalies (les 7 restantes
sont des écarts positifs plausibles, ex. pièces de rab).

Corrigé dans coupe_reader.py (+ main_window.py pour transmettre les
quantités attendues, models.py pour le type de qte_trouvee) et dans
le portage JS (coupe_controle/web/index.html).

+++ coupe_controle/models.py
@@ -33,7 +33,7 @@ class LigneComparaison:
     article: str
     lancement: str
     qte_attendue: float
-    qte_trouvee: int
+    qte_trouvee: float
     statut: Statut
 
     @property
